refactor(client): moving github files to folder
