+++ packages/amplication-client/src/Application/github/dialogs/GithubRepos.scss
@@ -1,4 +1,4 @@
-@import "../style/index.scss";
+@import "../../../style/index.scss";
 
 $max-height: 80vh;
 .github-repos {
